Move download and screenshot files to AppBar
